move tracebackBag, errorTxt, errorLog from gnrlang to gnrlog
Eliminates circular dependencies in gnr.core.gnrlang by moving
traceback/error functions to gnr.core.gnrlog where they belong
semantically. All runtime imports in gnrlang are now removed.

Refs #710
